StudentAdmissionPrediction
This is a simple implementation of a neural networks with forward path implementation to predict if a student gets admit in a particular university based on his GRE, GPA and rank. The implementation does not do backpropagation. It simply calculates the output and loss and has a prediction accuracy of 67% and training loss of 20%.